+++ src/apis/search/index.ts
@@ -1,5 +1,19 @@
 import { productItemType } from '@/types/Product/productList';
 import instance from '../axios';
+import requests from '../requests';
+
+export async function getRecentSearchList() {
+    try {
+        const response = await instance.get(requests.recentSearchList);
+        if (response.data) return response.data.list;
+        throw new Error(
+            `Unexpected response : ${response.status} ${response.statusText}`,
+        );
+    } catch (error) {
+        console.error(error);
+        return null;
+    }
+}
 
 export async function getQuerySearch(
     keyword: string,
@@ -23,5 +37,20 @@ export async function getQuerySearch(
     } catch (error) {
         console.error(error);
         throw Error;
+    }
+}
+
+export async function deleteRecentSearch(index: number) {
+    try {
+        const response = await instance.delete(
+            `${requests.recentSearchList}?index=${index}`,
+        );
+        if (response.data) return response.data.list;
+        throw new Error(
+            `Unexpected response : ${response.status} ${response.statusText}`,
+        );
+    } catch (error) {
+        console.error(error);
+        return null;
     }
 }
