+++ src/components/Product/CategoryList.tsx
@@ -1,5 +1,5 @@
-import useGetProductList from '@/hooks/query/product/useGetProductList';
 import { useParams, useSearchParams } from 'react-router-dom';
+import useGetProductList from '@/hooks/query/product';
 import ProductList from './ProductList';
 
 export default function CategoryList() {
