+++ src/components/common/index.tsx
@@ -1,3 +1,2 @@
 export { default as TopHeader } from './TopHeader';
 export { default as LikeBtn } from './LikeBtn';
-export { default as BottomSheet } from './BottomSheet';
