--- src/hooks/query/search/useGetRecentSearchList.ts
@@ -1,11 +1,0 @@
-import { getRecentSearchList } from '@/apis/search/getRecentSearchList';
-import { useQuery } from '@tanstack/react-query';
-
-export default function useGetRecentSearchList() {
-    const { data: recentSearchList, isLoading: isLoadingRecentSearchList } =
-        useQuery({
-            queryKey: ['recentSearchList'],
-            queryFn: getRecentSearchList,
-        });
-    return { recentSearchList, isLoadingRecentSearchList };
-}
