+++ src/pages/Inquiry/index.tsx
@@ -1,0 +1,2 @@
+export { default as InquiryListPage } from './InquiryListPage';
+export { default as InquiryPage } from './InquiryPage';
