+++ src/pages/Member/index.tsx
@@ -1,0 +1,5 @@
+export { default as SocialLoginPage } from './SocialLoginPage';
+export { default as SocialLoginRedirectPage } from './SocialLoginRedirectPage';
+
+export { default as RegisterPage } from './RegisterPage';
+export { default as RegisterSuccessPage } from './RegisterSuccessPage';
