+++ src/pages/MyPage/index.tsx
@@ -1,0 +1,5 @@
+export { default as MyPage } from './MyPage';
+export { default as MyLikePage } from './MyLikePage';
+export { default as MyInformationPage } from './MyInformationPage';
+export { default as EditInformationPage } from './EditInformationPage';
+export { default as EditAddressPage } from './EditAddressPage';
