+++ src/pages/Product/index.tsx
@@ -1,0 +1,2 @@
+export { default as ProductDetailPage } from './ProductDetailPage';
+export { default as ProductPage } from './ProductPage';
