+++ src/pages/Review/index.tsx
@@ -1,0 +1,1 @@
+export { default as SubmitReviewPage } from './SubmitReviewPage';
