+++ src/pages/Search/index.tsx
@@ -1,0 +1,1 @@
+export { default as SearchPage } from './SearchPage';
