--- src/router/login/LoginRouter.tsx
@@ -1,13 +1,0 @@
-import SocialLoginPage from '@/pages/Login/SocialLoginPage';
-import SocialLoginRedirectPage from '@/pages/Login/SocialLoginRedirectPage';
-
-export const LoginRouter = [
-    {
-        path: '/login',
-        element: <SocialLoginPage />,
-    },
-    {
-        path: '/authorize',
-        element: <SocialLoginRedirectPage />,
-    },
-];
