--- src/router/review/reviewRouter.tsx
@@ -1,13 +1,0 @@
-import SubmitReviewPage from '@/pages/Review/SubmitReviewPage';
-
-export const reviewRouter = [
-    {
-        path: '/review',
-        children: [
-            {
-                path: 'submit',
-                element: <SubmitReviewPage />,
-            },
-        ],
-    },
-];
