--- src/router/search/SearchRouter.tsx
@@ -1,8 +1,0 @@
-import SearchPage from '@/pages/Search/SearchPage';
-
-export const SearchRouter = [
-    {
-        path: '/search',
-        element: <SearchPage />,
-    },
-];
